Retirando camada de validações da Model e adicionando uma camada de services para ficar responsável dessa parte
